refactor: relocate and rename widget-form-store
