+++ apps/web/src/services/dashboards/shared/dashboard-widget-input-form/composables/use-widget-more-options.ts
@@ -9,8 +9,8 @@ import {
 import {
     getRefinedWidgetOptionsSchema,
 } from '@/services/dashboards/shared/dashboard-widget-input-form/helpers/schema-helper';
+import { useWidgetFormStore } from '@/services/dashboards/shared/dashboard-widget-input-form/widget-form-store';
 import { useDashboardDetailInfoStore } from '@/services/dashboards/store/dashboard-detail-info';
-import { useWidgetFormStore } from '@/services/dashboards/store/widget-form';
 import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
 
 interface UseWidgetMoreOptionsState {
